saving new hook method to seperate busness from  Ui layer
